move browser launch logic into src/recorder
