+++ src/recorder/types.ts
@@ -1,0 +1,11 @@
+export interface LaunchBrowserOptions {
+  url?: string
+  capture: {
+    browser: boolean
+  }
+}
+
+export type LaunchBrowserErrorReason =
+  | 'websocket-server-error'
+  | 'browser-launch'
+  | 'unknown'
