+++ src/views/Recorder/EmptyState.tsx
@@ -26,11 +26,11 @@ import { z } from 'zod'
 import { FieldGroup } from '@/components/Form'
 import { ProxyHealthWarning } from '@/components/ProxyHealthWarning'
 import { TextButton } from '@/components/TextButton'
-import { LaunchBrowserOptions } from '@/handlers/browser/types'
 import { useProxyHealthCheck } from '@/hooks/useProxyHealthCheck'
 import { useProxyStatus } from '@/hooks/useProxyStatus'
 import { useRecentURLs } from '@/hooks/useRecentURLs'
 import { useBrowserCheck, useSettings } from '@/hooks/useSettings'
+import { LaunchBrowserOptions } from '@/recorder/types'
 import { useStudioUIStore } from '@/store/ui'
 import { ProxyStatus } from '@/types'
